+++ sdk-core/src/test/java/care/data4life/sdk/CryptoServiceTest.java
@@ -23,6 +23,7 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
+import org.mockito.ArgumentCaptor;
 import org.mockito.Mockito;
 
 import java.io.IOException;
@@ -34,6 +35,8 @@ import java.security.NoSuchProviderException;
 import java.security.PrivateKey;
 import java.security.PublicKey;
 import java.security.SecureRandom;
+import java.security.spec.InvalidKeySpecException;
+import java.security.spec.PKCS8EncodedKeySpec;
 import java.util.Objects;
 
 import javax.crypto.BadPaddingException;
@@ -61,6 +64,7 @@ import care.data4life.sdk.test.util.TestSchedulerRule;
 import care.data4life.sdk.util.Base64;
 import io.reactivex.observers.TestObserver;
 
+import static org.junit.Assert.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
@@ -260,6 +264,75 @@ public class CryptoServiceTest {
         // then
         testSubscriber
                 .onComplete();
+    }
+
+    @Test
+    public void setGCKeyPairFromPemPrivateKey_shouldStoreCorrectPrivateKey() throws NoSuchAlgorithmException, InvalidKeySpecException {
+        // given
+        String base64TestKey = "MIIEowIBAAKCAQEArosQh7LO8HhpJxeHBwQpi12uxOKQsYYBzMOUeoerijGOwtQG\n" +
+                "sz656ZmQJNR62BIK2mHSSjIcpmu3ydE5fJkUxtF+dpTdmPtZ2URGEyt3/dXFe5RR\n" +
+                "i9hiIVwgWzjZAiVrWfIx4MtPEk9fMV2WKMLOa+o2ZEWqDLfiDjU8ixSaUc/Vtd5K\n" +
+                "msSun587+iBPTR33pGAG1u3If3GSkQPKkW3elRNLxL4twSL45+pmSIgkcTIrxo71\n" +
+                "0xqN5hWRr0dv19RSYHapJdgIPXDQkJiKULpF3/OWst8/OjtBhP3G8ne0FDoP4xId\n" +
+                "RDXCUat1R6MuJVDx+sr/GkHjFlMjKjFUHrXQqQIDAQABAoIBAAHQTh6q2/2hsq4G\n" +
+                "T4/iGjBpi8xd8lT16IThL2TKjhzEgRBDNcKdDz9/KgFH9/LQ1S4JwC6nMKcGDYXa\n" +
+                "V7eUu6OJP8ApsdfKHNfmHrhKRlfr5b5v/xzt5a8lDu0DvTWJgAESRDRqyGqPSpTv\n" +
+                "vQS1aYGzkFcgZjD1pDKzmOp1D1l0RAZdGxa5cTsZiovmX27nOg16qKdmUIRbMyQC\n" +
+                "CQqnX+uWvfAJrxmu+DUQthsVfm599TvZkuqBCCdUvri7dEFLPrV08kn7/ZnqlBf9\n" +
+                "ms/OQjCGI+41LXHyxrriG8noBlklQ/QrXe8kZPFAv0YWCTYEo2pDuowIZrByDBZc\n" +
+                "d8KM5nUCgYEA19rwq+PS0UM2reVkM3JHMXxZrxp5yADjUUYfsiMom0BRq4T8HBao\n" +
+                "eD5wMiTBjRhLuDzwnFF/o38Z/DFgsyhIjNlI3mdrQV1iroiHK/QDqkKk53xQofFl\n" +
+                "Aro7JtkyVMcT0htjZzVDobBEzH4c/BcstYrlXt/r27GMeH+D6nZaAJcCgYEAzwE5\n" +
+                "AmWMAf9EOVVsSf0s4DSJ3So7pUUlkCg9i5QU7gLu6gMAa19EyZO0aBflmTtuovm2\n" +
+                "302/giL6z/SgEbzkHHomvr2RZYwocYMbN30EDHi4JX8RkHS8bYioz3Vw4g6ELvSs\n" +
+                "CnLRA9FCX4vB2tJtYI/LwiZukWnDFUYZQW13oL8CgYBhLp9gpEfME1jQ3hBI4VCQ\n" +
+                "RQ4TufXOSCgP9WRbzVyA2WprsInZE5Jx4Jqe2NGTdrbQkg86Ma8nqxfF5W1F/AL9\n" +
+                "9u3JxAIUAblmHu3MqiXkR/D6j4u1/Xqeyb3L9cmlRaP02oPceayjZTr0XmsqTDzC\n" +
+                "13ABUQtddAhsT+zSaMqIrQKBgQCGAaqgTJC4ckH+Q7iYpVc5xYlCLabzNLI+gm5l\n" +
+                "P3XVJvz3bP4GhGQJgp8Vi/LModbbloC2SqShYHexzBEbqoaZkNIoRJwtevBrm44w\n" +
+                "+7N1R2kejQYX2BprZj6yHrr2/KLBqw78rJt2ty8an2Tdfb/k9PHZO/v0Et2BliGf\n" +
+                "Y3hADQKBgCdHaK1A3cxTQK4+EE3DsUM4ZnB+vUS78IiHTSpJvUY/gX8RcWxTJVvc\n" +
+                "nnxSlrzV/rxA7uI40DFiPFGVrVoXW1w0C1ASeL3siqv1aoZ5QiuCUv6ULKrbNBp7\n" +
+                "QyhWfy6A4sU7XdwJfNhQUAoLvvRrECtqMR7Ayn7xoWgeuhqQCHeg\n";
+
+        String pemTestKeyString = "-----BEGIN RSA PRIVATE KEY-----\n" +
+                base64TestKey +
+                "-----END RSA PRIVATE KEY-----\n";
+
+        // when
+        cryptoService.setGCKeyPairFromPemPrivateKey(pemTestKeyString);
+
+        //then
+        /* The base64 serialized input and out keys may differ due to irrelevant meta-data
+        differences and hence cannot be compared directly. Hence, we convert both the original
+        and the stored key to Java key objects before comparing them. */
+        final ArgumentCaptor<GCKeyPair> keyPairArg = ArgumentCaptor.forClass(GCKeyPair.class);
+        verify(mockStorage).storeKey(eq(PREFIX + GC_KEYPAIR), keyPairArg.capture());
+        final GCRSAKeyAlgorithm algorithm = new GCRSAKeyAlgorithm();
+        final java.security.KeyFactory keyFactory = java.security.KeyFactory.getInstance(algorithm.getCipher());
+
+        final String storedPrivateKeyBase64 = keyPairArg.getValue().getPrivateKeyBase64();
+        final PrivateKey storedJavaKey = getPrivateJavaKey(keyFactory, storedPrivateKeyBase64);
+
+        final String testKeyNoLinebreaksBase64 = base64TestKey.replace("\n", "");
+        final PrivateKey testJavaKey = getPrivateJavaKey(keyFactory, testKeyNoLinebreaksBase64);
+
+        assertEquals(testJavaKey, storedJavaKey);
+    }
+
+    /**
+     * Convert a base 64 DER format key into a java.sercurity PrivateKey object
+     *
+     * @param keyFactory        java.security.KeyFactory
+     * @param privateKeyBase64  Private key as base 64 string
+     * @return                  Key
+     * @throws InvalidKeySpecException
+     */
+    private PrivateKey getPrivateJavaKey(java.security.KeyFactory keyFactory, String privateKeyBase64) throws InvalidKeySpecException {
+        final String base64StoredPrivateKey = privateKeyBase64;
+        final byte[] storedPrivateKey = Base64.INSTANCE.decode(base64StoredPrivateKey);
+        final PKCS8EncodedKeySpec encodedStoredKeySpec = new PKCS8EncodedKeySpec(storedPrivateKey);
+        return keyFactory.generatePrivate(encodedStoredKeySpec);
     }
 
     @Test